+++ GameLib.Plugin/GameLib.Plugin.Gog/GogLauncher.cs
@@ -37,6 +37,8 @@ public class GogLauncher : ILauncher
 
     public string Executable { get; private set; } = string.Empty;
 
+    public Icon? ExecutableIcon => PathUtil.GetFileIcon(ExecutablePath);
+
     public IEnumerable<IGame> Games { get; private set; } = Enumerable.Empty<IGame>();
 
     public void Refresh(CancellationToken cancellationToken = default)