+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+
+
+## [1.1.2] - 2022-01-09
 ### Added
 - Add ExecutableIcon property on IGame interface / implementations
 - Add ExecutableIcon property on ILauncher interface / implementations
+- Add WPF GUI demo application
 
 
 ## [1.1.1] - 2022-27-08